% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord2map_label_tw.R
\name{coord2mapLabel}
\alias{coord2mapLabel}
\title{Get Map Information from Geographic Coordinates}
\usage{
coord2mapLabel(long = 120.5, lat = 23.45, scale = c(5, 25, 50, 100))
}
\arguments{
\item{long}{numeric. The longtitute.}

\item{lat}{numeric. The latitute.}

\item{scale}{integer. One of \code{5}, \code{25}, \code{50},
and \code{100}, corresponding to map size of \code{1/5,000},
\code{1/25,000}, \code{50,000}, and \code{100,000}, respectively.}
}
\value{
A list of map information.
 \describe{
  \item{\code{mapLabel}}{Map Label (Taiwan)}
  \item{\code{long_range}}{Longitudinal range covered by the map.}
  \item{\code{lat_range}}{Latitudinal range covered by the map.}
 }
}
\description{
Get Map Information from Geographic Coordinates
}
