% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{deci2deg}
\alias{deci2deg}
\title{Decimal to Degree-Minute-Second}
\usage{
deci2deg(x)
}
\arguments{
\item{x}{numeric vector. A list of numeric values in degree
to be converted to degree-minute-second.}
}
\value{
A list of numeric vectors with length 3.
}
\description{
Decimal to Degree-Minute-Second
}
