% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{deg2deci}
\alias{deg2deci}
\title{Degree-Minute-Second to Decimal}
\usage{
deg2deci(degree = 0, minute = 0, second = 0)
}
\arguments{
\item{degree}{numeric.}

\item{minute}{numeric. Must be between \code{0} and
\code{60}.}

\item{second}{numeric. Must be between \code{0} and
\code{60}.}
}
\value{
A single value that contains decimal.
}
\description{
Degree-Minute-Second to Decimal
}
