% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{deg2rad}
\alias{deg2rad}
\title{Degree to r.a.d}
\usage{
deg2rad(degree = 180)
}
\arguments{
\item{degree}{numeric vector. A list of numeric values in degree
to be converted to r.a.d.}
}
\value{
A numeric vector of length equal to the input.
}
\description{
Degree to r.a.d
}
