% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{rad2deg}
\alias{rad2deg}
\title{r.a.d to Degree (in decimal)}
\usage{
rad2deg(rad = pi)
}
\arguments{
\item{rad}{numeric vector. A list of numeric values in r.a.d
to be converted to degree.}
}
\value{
A numeric vector of length equal to the input.
}
\description{
r.a.d to Degree (in decimal)
}
